first commit, move VIEWTYPES enum from file_tab.h to custom.h, so it is accessible for other units
